pyproject.toml
[tool.poetry]
name = "lab-32"
version = "0.1.0"
description = ""
authors = ["noureddein <noureddein@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
